meter for happiness WORKS
it drains and fills back up! should be easy enough to do now

+++ PigeonNest/Assets/Scripts/ClickInteraction.cs
@@ -8,7 +8,7 @@ public class ClickInteraction : MonoBehaviour
     public HappyBar happyBar;
     public Pigeon pigeon;
 
-    public int happyGain;
+    public float happyGain;
 
     private void OnMouseDown()
     {
